[ stm32 port ] renaming BSP code to have STM32FX / stm32fx notation instead of F4 since it is more generic than specific
