+++ make/mt-config/mt-presets.mk
@@ -20,8 +20,8 @@ CONFIG_DUMMY_MIN                =memory_fs
 CONFIG_CC3200                   =memory_fs-tls_bsp
 CONFIG_CC3200_TLS_SOCKET        =memory_fs-tls_socket
 
-CONFIG_STM32F4                  =memory_fs-tls_bsp
-CONFIG_STM32F4_NUCLEO_WIFI      =memory_fs-tls_socket
+CONFIG_STM32FX                  =memory_fs-tls_bsp
+CONFIG_STM32FX_NUCLEO_WIFI      =memory_fs-tls_socket
 
 # TARGET presets
 TARGET_STATIC_DEV               =-static-debug
@@ -99,19 +99,19 @@ else ifeq ($(PRESET), CC3200_TLS_SOCKET)
     XI_TARGET_PLATFORM = cc3200
 
 # -------------------------------------------------------
-# ST Micro STM32F4
-else ifeq ($(PRESET), STM32F4)
-    CONFIG = $(CONFIG_STM32F4)
+# ST Micro STM32FX
+else ifeq ($(PRESET), STM32FX)
+    CONFIG = $(CONFIG_STM32FX)
     TARGET = $(TARGET_STATIC_REL)
-    XI_BSP_PLATFORM = stm32f4
-    XI_TARGET_PLATFORM = stm32f4
+    XI_BSP_PLATFORM = stm32fx
+    XI_TARGET_PLATFORM = stm32fx
 
-# ST Micro STM32F4_NUCLEO_WIFI
-else ifeq ($(PRESET), STM32F4_NUCLEO_WIFI)
-    CONFIG = $(CONFIG_STM32F4_NUCLEO_WIFI)
+# ST Micro STM32FX_NUCLEO_WIFI
+else ifeq ($(PRESET), STM32FX_NUCLEO_WIFI)
+    CONFIG = $(CONFIG_STM32FX_NUCLEO_WIFI)
     TARGET = $(TARGET_STATIC_REL)
-    XI_BSP_PLATFORM = stm32f4_nucleo_wifi
-    XI_TARGET_PLATFORM = stm32f4_nucleo_wifi
+    XI_BSP_PLATFORM = stm32fx_nucleo_wifi
+    XI_TARGET_PLATFORM = stm32fx_nucleo_wifi
 
 # -------------------------------------------------------
 # Fuzz Tests
