+++ make/mt-os/mt-os.mk
@@ -10,8 +10,8 @@ XI_CONST_PLATFORM_ARM := arm-linux
 XI_CONST_PLATFORM_WMSDK := wmsdk
 XI_CONST_PLATFORM_MICROCHIP := microchip
 XI_CONST_PLATFORM_CC3200 := cc3200
-XI_CONST_PLATFORM_STM32F4 := stm32f4
-XI_CONST_PLATFORM_STM32F4_NUCLEO_WIFI := stm32f4_nucleo_wifi
+XI_CONST_PLATFORM_STM32FX := stm32fx
+XI_CONST_PLATFORM_STM32FX_NUCLEO_WIFI := stm32fx_nucleo_wifi
 
 ifneq (,$(findstring $(XI_CONST_PLATFORM_ARM),$(TARGET)))
   XI_CONST_PLATFORM_CURRENT := $(XI_CONST_PLATFORM_ARM)
@@ -42,12 +42,12 @@ ifneq (,$(findstring $(XI_CONST_PLATFORM_CC3200),$(TARGET)))
   XI_CONST_PLATFORM_CURRENT := $(XI_CONST_PLATFORM_CC3200)
 endif
 
-ifneq (,$(findstring $(XI_CONST_PLATFORM_STM32F4),$(TARGET)))
-  XI_CONST_PLATFORM_CURRENT := $(XI_CONST_PLATFORM_STM32F4)
+ifneq (,$(findstring $(XI_CONST_PLATFORM_STM32FX),$(TARGET)))
+  XI_CONST_PLATFORM_CURRENT := $(XI_CONST_PLATFORM_STM32FX)
 endif
 
-ifneq (,$(findstring $(XI_CONST_PLATFORM_STM32F4_NUCLEO_WIFI),$(TARGET)))
-  XI_CONST_PLATFORM_CURRENT := $(XI_CONST_PLATFORM_STM32F4_NUCLEO_WIFI)
+ifneq (,$(findstring $(XI_CONST_PLATFORM_STM32FX_NUCLEO_WIFI),$(TARGET)))
+  XI_CONST_PLATFORM_CURRENT := $(XI_CONST_PLATFORM_STM32FX_NUCLEO_WIFI)
 endif
 
 XI_BINDIR := $(XI_BINDIR_BASE)/$(XI_CONST_PLATFORM_CURRENT)
