+++ src/bsp/platform/stm32fx/includes/lwipopts.h
