+++ src/bsp/platform/stm32fx/xi_bsp_rng_stm32fx.c
