+++ src/bsp/platform/stm32fx/xi_bsp_time_stm32fx.c
@@ -7,7 +7,7 @@
 #include <xi_bsp_time.h>
 
 #include <stm32f4xx_hal.h>
-#include <xi_bsp_time_stm32_sntp.h>
+#include <xi_bsp_time_stm32fx_sntp.h>
 
 void xi_bsp_time_init()
 {
